Mise a jour de la libSVM : 2.84 : mise à jour des fichiers svm.h et svm.cxx. Répercution sur la classe SMVModelEstimator : degree est désormais en int, donc plus de cast à faire -> plus propre.

+++ Code/Learning/otbSVMModelEstimator.h
@@ -120,14 +120,14 @@ public:
   /** Set the degree of the polynomial kernel */
   void SetPolynomialKernelDegree(int degree)
   {
-    param.degree = static_cast<double>(degree);
+    param.degree = degree;
   }
 
 
   /** Get the degree of the polynomial kernel */
   int GetPolynomialKernelDegree(void)
   {
-    return static_cast<int>(param.degree);
+    return param.degree;
   }
 
   /** Set the gamma parameter for poly/rbf/sigmoid kernels */
